Ready version

+++ Torrent/src/main/java/ru/nsu/yurakov/Torrent/ParserTorrent/TorrentInfo.java
@@ -3,7 +3,7 @@ package ru.nsu.yurakov.Torrent.ParserTorrent;
 public class TorrentInfo {
     public long sizeFile = 0;
     public String nameFile = null;
-    public int sizePiece = 0;
+    public long sizePiece = 0;
     public byte[] hashInfo = null;
     public byte[] pieces = null;
 }
